Update package structure
- Move actual test functions in v2 to shellwhat_ext/__init__.py
- Rename new test_cmd_line to test_cmd_line_v2 so they can exist next to each other
- Move all tests under the tests folder (which is now in root)
- Move README to docstring of test_cmd_line (v1)
