--- v2/README.md
@@ -1,3 +1,0 @@
-A simpler SCT function to test user command lines in the shell.
-
-Run tests with `python -m pytest` in this directory.
